backend-flask/requirement.txt
flask
numpy
Pillow
tensorflow





